refactor(page/bond): refactor and correct ui
